requirements.txt
# Core
flask>=2.3.0
python-dotenv>=1.0.0

# AWS
boto3>=1.28.0

# Azure OpenAI + LangGraph
openai>=1.12.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0

# Azure Identity (Service Principal + PEM certificate auth)
azure-identity>=1.15.0

# Utilities
numpy>=1.24.0
